# src/gen/java/it/tccr/puppet/api/AccountConnectionCallbackApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package it.tccr.puppet.api;

import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
public interface AccountConnectionCallbackApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * GET /account-connection-callback : Notify that aggregation completed
     * Notify that an account aggregation completed on account aggregation provider side. 
     *
     * @param code  (required)
     * @param credentialsId  (required)
     * @return Aggregation successfully completed (status code 302)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/account-connection-callback"
    )
    default ResponseEntity<Void> notifyAggregationCompleted(
         @RequestParam(value = "code", required = true) String code,
         @RequestParam(value = "credentialsId", required = true) String credentialsId
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}